FastAPI-Proj/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
blinker==1.6.2
certifi==2023.7.22
click==8.1.6
colorama==0.4.6
fastapi==0.100.1
Flask==2.3.2
graphql-core==3.2.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
priority==2.0.0
pydantic==2.1.1
pydantic_core==2.4.0
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
strawberry-graphql==0.196.2
typing_extensions==4.7.1
uvicorn==0.23.1
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.6
wsproto==1.2.0
